Fix App.js
